notification-service/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
motor==3.6.0
pymongo==4.10.1

# Kafka
confluent-kafka==2.3.0

# Push Notifications (optional - mock for dev)
# Optional for production push: firebase-admin, apns2

# Auth
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
hypothesis==6.92.0

# Utilities
python-dotenv==1.0.0
